Move Classes to Queries Namespace
